Cambios registro usuario e iniciando sesión y tipo de datos sesión
