--- flutter_application_1/lib/Interfaz/InicioSesion/index.dart
@@ -1,9 +1,0 @@
-export 'home.dart';
-export 'iniciarSesion.dart';
-export 'inicioApp.dart';
-export 'registrarse1.dart';
-export 'registrarse2.dart';
-export 'registrarse3.dart';
-export 'confirmarRegistro.dart';
-export 'registrandoUsuario.dart';
-export 'errorRegistro.dart';
