+++ flutter_application_1/lib/Interfaz/InicioSesion_Registro/Estilo/animaciones.dart
@@ -29,8 +29,8 @@ class _AnimacionImagenState extends State<AnimacionImagen>
 
   @override
   void dispose() {
-    super.dispose();
     animationController.dispose();
+    super.dispose();
   }
 
   @override
