+++ flutter_application_1/lib/Interfaz/InicioSesion_Registro/Registro/registrandoUsuario.dart
@@ -8,10 +8,10 @@ import 'dart:async';
 
 import 'package:flutter_application_1/Api/api.dart';
 
-import 'package:flutter_application_1/Interfaz/InicioSesion/Estilo/index.dart';
+import 'package:flutter_application_1/Interfaz/InicioSesion_Registro/Estilo/index.dart';
 
 import 'package:flutter_application_1/Data_types/registro.dart';
-import 'package:flutter_application_1/Interfaz/InicioSesion/index.dart';
+import 'package:flutter_application_1/Interfaz/InicioSesion_Registro/index.dart';
 import 'package:flutter_application_1/Interfaz/Menu/home.dart';
 
 class RegistrandoUsuario extends StatefulWidget {
@@ -61,18 +61,14 @@ class _RegistrandoUsuario extends State<RegistrandoUsuario> {
 
   void _setTimer() {
     if (registrado) {
-      _timer1.cancel();
-      _timer2.cancel();
-      Navigator.push(
-        context,
-        MaterialPageRoute(builder: (context) => Menu()),
-      );
+      Navigator.pushAndRemoveUntil(
+          context,
+          MaterialPageRoute(builder: (context) => Menu()),
+          (Route<dynamic> route) => false);
     } else if (contador < 7) {
       contador++;
       _timer2 = Timer(const Duration(seconds: 1), _setTimer);
     } else {
-      _timer1.cancel();
-      _timer2.cancel();
       Navigator.push(
         context,
         MaterialPageRoute(builder: (context) => ErrorRegistro()),
@@ -82,39 +78,44 @@ class _RegistrandoUsuario extends State<RegistrandoUsuario> {
 
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
-      resizeToAvoidBottomInset: false,
-      body: Container(
-        // ignore: prefer_const_constructors
-        decoration: BoxDecoration(
-          image: const DecorationImage(
-              image: AssetImage('assets/tapete.png'), fit: BoxFit.fill),
-        ),
-        child: Align(
-          alignment: Alignment.center,
-          child: Stack(
-            children: const [
-              Padding(
-                padding: EdgeInsets.only(top: 15),
-                child: Text(
-                  "Registrando usuario...",
-                  style: TextStyle(
-                    fontSize: 30.0,
-                    fontWeight: FontWeight.bold,
-                    color: Color(0xFFf7f6f6),
-                    fontFamily: "Bona Nova",
+    return WillPopScope(
+      onWillPop: () async {
+        return false;
+      },
+      child: Scaffold(
+        resizeToAvoidBottomInset: false,
+        body: Container(
+          // ignore: prefer_const_constructors
+          decoration: BoxDecoration(
+            image: const DecorationImage(
+                image: AssetImage('assets/tapete.png'), fit: BoxFit.fill),
+          ),
+          child: Align(
+            alignment: Alignment.center,
+            child: Stack(
+              children: const [
+                Padding(
+                  padding: EdgeInsets.only(top: 15),
+                  child: Text(
+                    "Registrando usuario...",
+                    style: TextStyle(
+                      fontSize: 30.0,
+                      fontWeight: FontWeight.bold,
+                      color: Color(0xFFf7f6f6),
+                      fontFamily: "Bona Nova",
+                    ),
                   ),
                 ),
-              ),
-              Padding(
-                padding: EdgeInsets.only(left: 295),
-                child: SizedBox(
-                  width: 60,
-                  height: 60,
-                  child: AnimacionImagen(),
+                Padding(
+                  padding: EdgeInsets.only(left: 295),
+                  child: SizedBox(
+                    width: 60,
+                    height: 60,
+                    child: AnimacionImagen(),
+                  ),
                 ),
-              ),
-            ],
+              ],
+            ),
           ),
         ),
       ),
