+++ flutter_application_1/lib/Interfaz/InicioSesion_Registro/index.dart
@@ -1,0 +1,9 @@
+export 'inicioApp.dart';
+export 'home.dart';
+export 'InicioSesion/iniciarSesion.dart';
+export 'Registro/registrarse1.dart';
+export 'Registro/registrarse2.dart';
+export 'Registro/registrarse3.dart';
+export 'Registro/confirmarRegistro.dart';
+export 'Registro/registrandoUsuario.dart';
+export 'Registro/errorRegistro.dart';
